refactor(engine): Move the push session to its own file (no-changelog)
