+++ packages/cli/src/services/engine-v2-push-registry.service.ts
@@ -1,44 +1,7 @@
 import { Time } from '@n8n/constants';
 import { Service } from '@n8n/di';
-import type { INodeExecutionData } from 'n8n-workflow';
 
-/**
- * A step run reported to the editor. Kept after it settles so a redelivered
- * event is ignored instead of appending a duplicate.
- */
-export class EngineV2StepRun {
-	/** Whether the step's outcome has been reported. */
-	settled = false;
-
-	constructor(
-		/** Pairs this run's `nodeExecuteAfter` with its `nodeExecuteAfterData`. */
-		readonly executionIndex: number,
-		readonly startTime: number,
-	) {}
-}
-
-/** State needed to relay one execution's lifecycle events to the editor. */
-export class EngineV2PushSession {
-	/** Ordering counter for `nodeExecuteBefore`/`nodeExecuteAfter`; starts at 0. */
-	sequenceNumber = 0;
-	/** Next `ITaskData.executionIndex` to hand out. */
-	nextExecutionIndex = 0;
-	/** Step runs keyed by the engine's step id. */
-	readonly steps = new Map<string, EngineV2StepRun>();
-	/** When the last lifecycle event for this execution arrived. */
-	lastSeenAt = Date.now();
-
-	constructor(
-		/** The only routing key {@link Push.send} accepts. */
-		readonly pushRef: string,
-		readonly workflowId: string,
-		/**
-		 * The trigger's outputs, since the engine never announces it as a step.
-		 * Cleared once emitted — pinned data can be large.
-		 */
-		public trigger?: { nodeName: string; outputs: INodeExecutionData[][] },
-	) {}
-}
+import { EngineV2PushSession } from '@/services/engine-v2-push-session';
 
 /** Long enough to outlive a run that idles between steps, e.g. on a wait. */
 const SESSION_TTL_MS = 12 * Time.hours.toMilliseconds;
@@ -54,7 +17,6 @@ const MAX_SESSIONS = 500;
  */
 @Service()
 export class EngineV2PushRegistry {
-	/** Ordered least recently seen first, so eviction reads from the front. */
 	private readonly sessions = new Map<string, EngineV2PushSession>();
 
 	register(
@@ -70,12 +32,7 @@ export class EngineV2PushRegistry {
 
 	get(executionId: string): EngineV2PushSession | undefined {
 		const session = this.sessions.get(executionId);
-		if (!session) return undefined;
-
-		session.lastSeenAt = Date.now();
-		// Re-insert to move the session to the back of the eviction order.
-		this.sessions.delete(executionId);
-		this.sessions.set(executionId, session);
+		if (session) session.lastSeenAt = Date.now();
 
 		return session;
 	}
@@ -92,15 +49,16 @@ export class EngineV2PushRegistry {
 	private evict(): void {
 		const cutoff = Date.now() - SESSION_TTL_MS;
 		for (const [executionId, session] of this.sessions) {
-			if (session.lastSeenAt >= cutoff) break; // rest are newer
-			this.sessions.delete(executionId);
+			if (session.lastSeenAt < cutoff) this.sessions.delete(executionId);
 		}
 
 		// Leave room for the caller's session, so the cap holds after the insert.
-		while (this.sessions.size >= MAX_SESSIONS) {
-			const oldest = this.sessions.keys().next().value;
-			if (oldest === undefined) break;
-			this.sessions.delete(oldest);
+		const excess = this.sessions.size - MAX_SESSIONS + 1;
+		if (excess <= 0) return;
+
+		const oldestFirst = [...this.sessions].sort((a, b) => a[1].lastSeenAt - b[1].lastSeenAt);
+		for (const [executionId] of oldestFirst.slice(0, excess)) {
+			this.sessions.delete(executionId);
 		}
 	}
 }
